refactor(runtime): extract approval gate and call prep into agent/turn — #7415
Pair commit (tightly coupled): the per-call prep loop moves to
turn/call_prep.rs::prepare_tool_calls and its approval section to
turn/approval_gate.rs::gate_tool_approval -> ApprovalGateOutcome
{Proceed{approved}, Deny(..), Replace(..)}. Deny/Replace carry the
synthesized outcome; the caller records it into ordered_results.

seen_tool_signatures stays owned by the loop skeleton (reset per
iteration) and is passed &mut into prepare_tool_calls, per the
RUN_SHEET dedup-state placement rule.

Import hygiene (clippy -D warnings): loop_.rs drops the now-unused
ApprovalRequest/ApprovalRequirement/ApprovalResponse imports; the
maybe_inject_channel_delivery_defaults re-export (loop_.rs and
turn/mod.rs) is now #[cfg(test)] — its only lib user moved into
turn/call_prep.rs; loop_.rs tests still import it.
