+++ crates/zeroclaw-runtime/src/agent/loop_.rs
@@ -1,4 +1,4 @@
-use crate::approval::{ApprovalManager, ApprovalRequest, ApprovalRequirement, ApprovalResponse};
+use crate::approval::ApprovalManager;
 
 /// CLI channel factory, injected by the binary. Returns a `Box<dyn Channel>` for interactive mode.
 pub static CLI_CHANNEL_FN: std::sync::OnceLock<
@@ -638,6 +638,7 @@ pub async fn agent_turn(
     .await
 }
 
+#[cfg(test)]
 pub(crate) use super::turn::maybe_inject_channel_delivery_defaults;
 
 // ── Agent Tool-Call Loop ──────────────────────────────────────────────────
@@ -1079,333 +1080,19 @@ pub async fn run_tool_call_loop(
         // tool executions concurrently for lower wall-clock latency.
         let mut tool_results = String::new();
         let mut individual_results: Vec<(Option<String>, String)> = Vec::new();
-        let mut ordered_results: Vec<Option<(String, Option<String>, ToolExecutionOutcome)>> =
-            (0..tool_calls.len()).map(|_| None).collect();
         let allow_parallel_execution =
             parallel_tools && should_execute_tools_in_parallel(&tool_calls, approval);
-        let mut executable_indices: Vec<usize> = Vec::new();
-        let mut executable_calls: Vec<ParsedToolCall> = Vec::new();
-
-        for (idx, call) in tool_calls.iter().enumerate() {
-            // ── Hook: before_tool_call (modifying) ──────────
-            let mut tool_name = call.name.clone();
-            let mut tool_args = call.arguments.clone();
-            if let Some(hooks) = hooks {
-                match hooks
-                    .run_before_tool_call(tool_name.clone(), tool_args.clone())
-                    .await
-                {
-                    crate::hooks::HookResult::Cancel(reason) => {
-                        ::zeroclaw_log::record!(INFO, ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Note).with_attrs(::serde_json::json!({"tool": call.name, "reason": reason.to_string()})), "tool call cancelled by hook");
-                        let cancelled = format!("Cancelled by hook: {reason}");
-                        ::zeroclaw_log::record!(
-                            WARN,
-                            ::zeroclaw_log::Event::new(
-                                module_path!(),
-                                ::zeroclaw_log::Action::Cancel
-                            )
-                            .with_outcome(::zeroclaw_log::EventOutcome::Failure)
-                            .with_attrs(::serde_json::json!({
-                                "model": model,
-                                "iteration": iteration + 1,
-                                "tool": call.name,
-                                "arguments": scrub_credentials(&tool_args.to_string()),
-                                "result": cancelled,
-                                "trace_id": turn_id,
-                            })),
-                            "tool_call_result"
-                        );
-                        if let Some(ref tx) = on_delta {
-                            let _ = tx
-                                .send(StreamDelta::Status(format!(
-                                    "\u{274c} {}: {}\n",
-                                    call.name,
-                                    truncate_with_ellipsis(&scrub_credentials(&cancelled), 200)
-                                )))
-                                .await;
-                        }
-                        ordered_results[idx] = Some((
-                            call.name.clone(),
-                            call.tool_call_id.clone(),
-                            ToolExecutionOutcome {
-                                output: cancelled,
-                                success: false,
-                                error_reason: Some(scrub_credentials(&reason)),
-                                duration: Duration::ZERO,
-                                receipt: None,
-                            },
-                        ));
-                        continue;
-                    }
-                    crate::hooks::HookResult::Continue((name, args)) => {
-                        tool_name = name;
-                        tool_args = args;
-                    }
-                }
-            }
-
-            maybe_inject_channel_delivery_defaults(
-                &tool_name,
-                &mut tool_args,
-                channel_name,
-                channel_reply_target,
-            );
-
-            super::set_runtime_approved_arg(&tool_name, &mut tool_args, false);
-
-            // ── Approval hook ────────────────────────────────
-            let mut approval_requirement = approval
-                .map(|mgr| mgr.approval_requirement(&tool_name))
-                .unwrap_or(ApprovalRequirement::NotRequired);
-            if let Some(mgr) = approval
-                && approval_requirement == ApprovalRequirement::Prompt
-            {
-                let request = ApprovalRequest {
-                    tool_name: tool_name.clone(),
-                    arguments: tool_args.clone(),
-                };
-
-                // Interactive CLI: prompt the operator.
-                // Non-interactive (channels): try the channel's inline
-                // approval (e.g. Telegram inline keyboard) before falling
-                // back to auto-deny.
-                let decision = if mgr.is_non_interactive() {
-                    let channel_decision = if let Some(ch) = channel {
-                        let ch_request = zeroclaw_api::channel::ChannelApprovalRequest {
-                            tool_name: request.tool_name.clone(),
-                            arguments_summary: crate::approval::summarize_args(&request.arguments),
-                            raw_arguments: Some(request.arguments.clone()),
-                        };
-                        let recipient = channel_reply_target.unwrap_or_default();
-                        match ch.request_approval(recipient, &ch_request).await {
-                            Ok(Some(r)) => Some(r),
-                            Ok(None) => None,
-                            Err(e) => {
-                                ::zeroclaw_log::record!(
-                                    WARN,
-                                    ::zeroclaw_log::Event::new(
-                                        module_path!(),
-                                        ::zeroclaw_log::Action::Note
-                                    )
-                                    .with_outcome(::zeroclaw_log::EventOutcome::Unknown)
-                                    .with_attrs(::serde_json::json!({"error": format!("{}", e)})),
-                                    "Channel approval request failed"
-                                );
-                                None
-                            }
-                        }
-                    } else {
-                        None
-                    };
-                    match channel_decision {
-                        Some(zeroclaw_api::channel::ChannelApprovalResponse::Approve) => {
-                            ApprovalResponse::Yes
-                        }
-                        Some(zeroclaw_api::channel::ChannelApprovalResponse::AlwaysApprove) => {
-                            ApprovalResponse::Always
-                        }
-                        Some(zeroclaw_api::channel::ChannelApprovalResponse::Deny) => {
-                            ApprovalResponse::No
-                        }
-                        Some(zeroclaw_api::channel::ChannelApprovalResponse::DenyWithEdit {
-                            replacement,
-                        }) => ApprovalResponse::ReplaceWith(replacement),
-                        // Channel doesn't support approval — auto-deny.
-                        None => ApprovalResponse::No,
-                    }
-                } else {
-                    mgr.prompt_cli(&request)
-                };
-
-                mgr.record_decision(&tool_name, &tool_args, &decision, channel_name);
-
-                if decision == ApprovalResponse::No {
-                    let denied = "Denied by user.".to_string();
-                    ::zeroclaw_log::record!(
-                        WARN,
-                        ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Reject)
-                            .with_outcome(::zeroclaw_log::EventOutcome::Failure)
-                            .with_attrs(::serde_json::json!({
-                                "model": model,
-                                "iteration": iteration + 1,
-                                "tool": tool_name.clone(),
-                                "arguments": scrub_credentials(&tool_args.to_string()),
-                                "result": denied,
-                                "trace_id": turn_id,
-                            })),
-                        "tool_call_result"
-                    );
-                    if let Some(ref tx) = on_delta {
-                        let _ = tx
-                            .send(StreamDelta::Status(format!(
-                                "\u{274c} {}: {}\n",
-                                tool_name, denied
-                            )))
-                            .await;
-                    }
-                    ordered_results[idx] = Some((
-                        tool_name.clone(),
-                        call.tool_call_id.clone(),
-                        ToolExecutionOutcome {
-                            output: denied.clone(),
-                            success: false,
-                            error_reason: Some(denied),
-                            duration: Duration::ZERO,
-                            receipt: None,
-                        },
-                    ));
-                    continue;
-                }
-
-                if let ApprovalResponse::ReplaceWith(replacement) = &decision {
-                    if let Some(ref tx) = on_delta {
-                        let _ = tx
-                            .send(StreamDelta::Status(format!(
-                                "\u{270f} {}: replaced by user\n",
-                                tool_name
-                            )))
-                            .await;
-                    }
-                    ::zeroclaw_log::record!(
-                        INFO,
-                        ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Note)
-                            .with_outcome(::zeroclaw_log::EventOutcome::Success)
-                            .with_attrs(::serde_json::json!({
-                                "model": model,
-                                "iteration": iteration + 1,
-                                "tool": tool_name.clone(),
-                                "arguments": scrub_credentials(&tool_args.to_string()),
-                                "replaced": true,
-                                "output": scrub_credentials(replacement),
-                                "trace_id": turn_id,
-                            })),
-                        "tool_call_result"
-                    );
-                    ordered_results[idx] = Some((
-                        tool_name.clone(),
-                        call.tool_call_id.clone(),
-                        ToolExecutionOutcome {
-                            output: crate::approval::sanitize_tool_replacement(replacement),
-                            success: true,
-                            error_reason: None,
-                            duration: Duration::ZERO,
-                            receipt: None,
-                        },
-                    ));
-                    continue;
-                }
-
-                if matches!(decision, ApprovalResponse::Yes | ApprovalResponse::Always) {
-                    approval_requirement = ApprovalRequirement::Approved;
-                }
-            }
-            super::set_runtime_approved_arg(
-                &tool_name,
-                &mut tool_args,
-                approval_requirement == ApprovalRequirement::Approved,
-            );
-
-            let signature = {
-                let canonical_args = canonicalize_json_for_tool_signature(&tool_args);
-                let args_json =
-                    serde_json::to_string(&canonical_args).unwrap_or_else(|_| "{}".to_string());
-                (tool_name.trim().to_ascii_lowercase(), args_json)
-            };
-            let dedup_exempt = dedup_exempt_tools.iter().any(|e| e == &tool_name)
-                || crate::tools::REENTRANT_AGENT_TOOLS.contains(&tool_name.as_str());
-            if !dedup_exempt && !seen_tool_signatures.insert(signature) {
-                let duplicate = format!(
-                    "Skipped duplicate tool call '{tool_name}' with identical arguments in this turn."
-                );
-                ::zeroclaw_log::record!(
-                    INFO,
-                    ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Skip)
-                        .with_outcome(::zeroclaw_log::EventOutcome::Failure)
-                        .with_attrs(::serde_json::json!({
-                            "model": model,
-                            "iteration": iteration + 1,
-                            "tool": tool_name.clone(),
-                            "arguments": scrub_credentials(&tool_args.to_string()),
-                            "result": duplicate,
-                            "deduplicated": true,
-                            "trace_id": turn_id,
-                        })),
-                    "tool_call_result"
-                );
-                if let Some(ref tx) = on_delta {
-                    let _ = tx
-                        .send(StreamDelta::Status(format!(
-                            "\u{274c} {}: {}\n",
-                            tool_name, duplicate
-                        )))
-                        .await;
-                }
-                ordered_results[idx] = Some((
-                    tool_name.clone(),
-                    call.tool_call_id.clone(),
-                    ToolExecutionOutcome {
-                        output: duplicate.clone(),
-                        success: false,
-                        error_reason: Some(duplicate),
-                        duration: Duration::ZERO,
-                        receipt: None,
-                    },
-                ));
-                continue;
-            }
-
-            ::zeroclaw_log::record!(
-                INFO,
-                ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Start)
-                    .with_attrs(::serde_json::json!({
-                        "model": model,
-                        "iteration": iteration + 1,
-                        "tool": tool_name.clone(),
-                        "arguments": scrub_credentials(&tool_args.to_string()),
-                        "trace_id": turn_id,
-                    })),
-                "tool_call_start"
-            );
-
-            // ── Progress: tool start ────────────────────────────
-            if let Some(ref tx) = on_delta {
-                let hint = {
-                    let raw = match tool_name.as_str() {
-                        "shell" => tool_args.get("command").and_then(|v| v.as_str()),
-                        "file_read" | "file_write" => {
-                            tool_args.get("path").and_then(|v| v.as_str())
-                        }
-                        _ => tool_args
-                            .get("action")
-                            .and_then(|v| v.as_str())
-                            .or_else(|| tool_args.get("query").and_then(|v| v.as_str())),
-                    };
-                    match raw {
-                        Some(s) => truncate_with_ellipsis(s, 60),
-                        None => String::new(),
-                    }
-                };
-                let progress = if hint.is_empty() {
-                    format!("\u{23f3} {}\n", tool_name)
-                } else {
-                    format!("\u{23f3} {}: {hint}\n", tool_name)
-                };
-                ::zeroclaw_log::record!(
-                    DEBUG,
-                    ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Note)
-                        .with_attrs(::serde_json::json!({"tool": tool_name})),
-                    "Sending progress start to draft"
-                );
-                let _ = tx.send(StreamDelta::Status(progress)).await;
-            }
-
-            executable_indices.push(idx);
-            executable_calls.push(ParsedToolCall {
-                name: tool_name,
-                arguments: tool_args,
-                tool_call_id: call.tool_call_id.clone(),
-            });
-        }
+        let super::turn::PreparedToolCalls {
+            mut ordered_results,
+            executable_indices,
+            executable_calls,
+        } = super::turn::prepare_tool_calls(
+            &ctx,
+            &tool_calls,
+            &mut seen_tool_signatures,
+            iteration,
+        )
+        .await;
 
         let executed_outcomes = if allow_parallel_execution && executable_calls.len() > 1 {
             execute_tools_parallel(
