+++ crates/zeroclaw-runtime/src/agent/turn/call_prep.rs
@@ -1,0 +1,223 @@
+//! The per-call preparation loop: `before_tool_call` hook, delivery defaults,
+//! the approval gate, the duplicate-call gate, and start logging — producing
+//! the executable subset of this round's tool calls.
+
+use super::approval_gate::{ApprovalGateOutcome, gate_tool_approval};
+use super::context::TurnCtx;
+use super::delivery_defaults::maybe_inject_channel_delivery_defaults;
+use super::events::StreamDelta;
+use super::redact::scrub_credentials;
+use crate::agent::tool_execution::ToolExecutionOutcome;
+use crate::util::truncate_with_ellipsis;
+use std::collections::HashSet;
+use std::time::Duration;
+use zeroclaw_tool_call_parser::{ParsedToolCall, canonicalize_json_for_tool_signature};
+
+/// The prepared subset of one round's tool calls.
+///
+/// `ordered_results` has one slot per incoming call; prep fills the slots for
+/// calls that never execute (hook-cancelled, denied, replaced, deduplicated)
+/// and post-exec fills the rest, so result ordering always matches the
+/// model's call ordering. `seen_tool_signatures` stays owned by the loop
+/// skeleton (reset per iteration) and is threaded in as `&mut`.
+pub(crate) struct PreparedToolCalls {
+    pub(crate) ordered_results: Vec<Option<(String, Option<String>, ToolExecutionOutcome)>>,
+    pub(crate) executable_indices: Vec<usize>,
+    pub(crate) executable_calls: Vec<ParsedToolCall>,
+}
+
+/// Run per-call preparation over this round's parsed tool calls (upstream
+/// loop body, per-call prep loop).
+pub(crate) async fn prepare_tool_calls(
+    ctx: &TurnCtx<'_>,
+    tool_calls: &[ParsedToolCall],
+    seen_tool_signatures: &mut HashSet<(String, String)>,
+    iteration: usize,
+) -> PreparedToolCalls {
+    let mut ordered_results: Vec<Option<(String, Option<String>, ToolExecutionOutcome)>> =
+        (0..tool_calls.len()).map(|_| None).collect();
+    let mut executable_indices: Vec<usize> = Vec::new();
+    let mut executable_calls: Vec<ParsedToolCall> = Vec::new();
+
+    for (idx, call) in tool_calls.iter().enumerate() {
+        // ── Hook: before_tool_call (modifying) ──────────
+        let mut tool_name = call.name.clone();
+        let mut tool_args = call.arguments.clone();
+        if let Some(hooks) = ctx.hooks {
+            match hooks
+                .run_before_tool_call(tool_name.clone(), tool_args.clone())
+                .await
+            {
+                crate::hooks::HookResult::Cancel(reason) => {
+                    ::zeroclaw_log::record!(INFO, ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Note).with_attrs(::serde_json::json!({"tool": call.name, "reason": reason.to_string()})), "tool call cancelled by hook");
+                    let cancelled = format!("Cancelled by hook: {reason}");
+                    ::zeroclaw_log::record!(
+                        WARN,
+                        ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Cancel)
+                            .with_outcome(::zeroclaw_log::EventOutcome::Failure)
+                            .with_attrs(::serde_json::json!({
+                                "model": ctx.model,
+                                "iteration": iteration + 1,
+                                "tool": call.name,
+                                "arguments": scrub_credentials(&tool_args.to_string()),
+                                "result": cancelled,
+                                "trace_id": ctx.turn_id,
+                            })),
+                        "tool_call_result"
+                    );
+                    if let Some(tx) = ctx.on_delta {
+                        let _ = tx
+                            .send(StreamDelta::Status(format!(
+                                "\u{274c} {}: {}\n",
+                                call.name,
+                                truncate_with_ellipsis(&scrub_credentials(&cancelled), 200)
+                            )))
+                            .await;
+                    }
+                    ordered_results[idx] = Some((
+                        call.name.clone(),
+                        call.tool_call_id.clone(),
+                        ToolExecutionOutcome {
+                            output: cancelled,
+                            success: false,
+                            error_reason: Some(scrub_credentials(&reason)),
+                            duration: Duration::ZERO,
+                            receipt: None,
+                        },
+                    ));
+                    continue;
+                }
+                crate::hooks::HookResult::Continue((name, args)) => {
+                    tool_name = name;
+                    tool_args = args;
+                }
+            }
+        }
+
+        maybe_inject_channel_delivery_defaults(
+            &tool_name,
+            &mut tool_args,
+            ctx.channel_name,
+            ctx.channel_reply_target,
+        );
+
+        crate::agent::set_runtime_approved_arg(&tool_name, &mut tool_args, false);
+
+        // ── Approval hook ────────────────────────────────
+        let approved = match gate_tool_approval(ctx, &tool_name, &tool_args, iteration).await {
+            ApprovalGateOutcome::Proceed { approved } => approved,
+            ApprovalGateOutcome::Deny(outcome) | ApprovalGateOutcome::Replace(outcome) => {
+                ordered_results[idx] =
+                    Some((tool_name.clone(), call.tool_call_id.clone(), outcome));
+                continue;
+            }
+        };
+        crate::agent::set_runtime_approved_arg(&tool_name, &mut tool_args, approved);
+
+        let signature = {
+            let canonical_args = canonicalize_json_for_tool_signature(&tool_args);
+            let args_json =
+                serde_json::to_string(&canonical_args).unwrap_or_else(|_| "{}".to_string());
+            (tool_name.trim().to_ascii_lowercase(), args_json)
+        };
+        let dedup_exempt = ctx.dedup_exempt_tools.iter().any(|e| e == &tool_name)
+            || crate::tools::REENTRANT_AGENT_TOOLS.contains(&tool_name.as_str());
+        if !dedup_exempt && !seen_tool_signatures.insert(signature) {
+            let duplicate = format!(
+                "Skipped duplicate tool call '{tool_name}' with identical arguments in this turn."
+            );
+            ::zeroclaw_log::record!(
+                INFO,
+                ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Skip)
+                    .with_outcome(::zeroclaw_log::EventOutcome::Failure)
+                    .with_attrs(::serde_json::json!({
+                        "model": ctx.model,
+                        "iteration": iteration + 1,
+                        "tool": tool_name.clone(),
+                        "arguments": scrub_credentials(&tool_args.to_string()),
+                        "result": duplicate,
+                        "deduplicated": true,
+                        "trace_id": ctx.turn_id,
+                    })),
+                "tool_call_result"
+            );
+            if let Some(tx) = ctx.on_delta {
+                let _ = tx
+                    .send(StreamDelta::Status(format!(
+                        "\u{274c} {}: {}\n",
+                        tool_name, duplicate
+                    )))
+                    .await;
+            }
+            ordered_results[idx] = Some((
+                tool_name.clone(),
+                call.tool_call_id.clone(),
+                ToolExecutionOutcome {
+                    output: duplicate.clone(),
+                    success: false,
+                    error_reason: Some(duplicate),
+                    duration: Duration::ZERO,
+                    receipt: None,
+                },
+            ));
+            continue;
+        }
+
+        ::zeroclaw_log::record!(
+            INFO,
+            ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Start).with_attrs(
+                ::serde_json::json!({
+                    "model": ctx.model,
+                    "iteration": iteration + 1,
+                    "tool": tool_name.clone(),
+                    "arguments": scrub_credentials(&tool_args.to_string()),
+                    "trace_id": ctx.turn_id,
+                })
+            ),
+            "tool_call_start"
+        );
+
+        // ── Progress: tool start ────────────────────────────
+        if let Some(tx) = ctx.on_delta {
+            let hint = {
+                let raw = match tool_name.as_str() {
+                    "shell" => tool_args.get("command").and_then(|v| v.as_str()),
+                    "file_read" | "file_write" => tool_args.get("path").and_then(|v| v.as_str()),
+                    _ => tool_args
+                        .get("action")
+                        .and_then(|v| v.as_str())
+                        .or_else(|| tool_args.get("query").and_then(|v| v.as_str())),
+                };
+                match raw {
+                    Some(s) => truncate_with_ellipsis(s, 60),
+                    None => String::new(),
+                }
+            };
+            let progress = if hint.is_empty() {
+                format!("\u{23f3} {}\n", tool_name)
+            } else {
+                format!("\u{23f3} {}: {hint}\n", tool_name)
+            };
+            ::zeroclaw_log::record!(
+                DEBUG,
+                ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Note)
+                    .with_attrs(::serde_json::json!({"tool": tool_name})),
+                "Sending progress start to draft"
+            );
+            let _ = tx.send(StreamDelta::Status(progress)).await;
+        }
+
+        executable_indices.push(idx);
+        executable_calls.push(ParsedToolCall {
+            name: tool_name,
+            arguments: tool_args,
+            tool_call_id: call.tool_call_id.clone(),
+        });
+    }
+
+    PreparedToolCalls {
+        ordered_results,
+        executable_indices,
+        executable_calls,
+    }
+}
