+++ crates/zeroclaw-runtime/src/agent/turn/mod.rs
@@ -5,6 +5,8 @@
 //! `/opt/notes/work/zeroclaw/unification_modular/RUN_SHEET.md` during the
 //! #7415 migration.
 
+pub(crate) mod approval_gate;
+pub(crate) mod call_prep;
 pub(crate) mod context;
 pub(crate) mod context_recovery;
 pub(crate) mod delivery_defaults;
@@ -20,8 +22,10 @@ pub(crate) mod stream_guard;
 pub(crate) mod tool_specs;
 pub(crate) mod vision_route;
 
+pub(crate) use call_prep::{PreparedToolCalls, prepare_tool_calls};
 pub(crate) use context::TurnCtx;
 pub(crate) use context_recovery::{record_llm_failure, try_recover_context_overflow};
+#[cfg(test)]
 pub(crate) use delivery_defaults::maybe_inject_channel_delivery_defaults;
 pub(crate) use events::STREAM_CHUNK_MIN_CHARS;
 pub use events::{DraftEvent, PROGRESS_MIN_INTERVAL_MS, StreamDelta};
